app/requirements.txt
absl-py==0.8.1
astor==0.8.0
beautifulsoup4==4.8.2
bleach==1.5.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
enum34==1.1.9
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
icrawler==0.6.2
idna==2.8
imageio==2.6.1
itsdangerous==1.1.0
jaconv==0.2.4
Jinja2==2.10.3
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mecab-python3==0.996.2
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydot==1.4.1
pyparsing==2.4.5
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.2
selenium==3.141.0
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
